Created npc_battle module, refactored code, and moved will_fight
