+++ include/npc/npc.h
@@ -3,46 +3,12 @@
 
 #include "game-state/game_state_common.h"
 #include "game-state/item.h"
-#include "game-state/stats.h"
-#include "npc/dialogue.h"
 #include "playerclass/class_structs.h"
 #include "playerclass/class.h"
-#include "battle/battle_common.h"
-#include "battle/battle_structs.h"
-#include "battle/battle_moves.h"
+#include "npc/dialogue.h"
+#include "npc/npc_battle.h"
 
 // NPC STRUCTURE DEFINITION ---------------------------------------------------
-
-/* Hostility level options */
-typedef enum hostility {
-    FRIENDLY = 0,
-    CONDITIONAL_FRIENDLY = 1,
-    HOSTILE = 2
-} hostility_t;
-
-/* Info used for battles with NPCs */
-typedef struct npc_battle {
-    /* NPC health level */
-    int health;
-
-    /* pointer to an existing stat struct */
-    stat_t *stats;
-
-    /* pointer to an existing move struct */
-    move_t *moves;
-
-    /* difficulty of the NPC's ai */
-    difficulty_t ai;
-    
-    /* hostility level of the npc */
-    hostility_t hostility_level;
-
-    /* whether the NPC will fight */
-    bool will_fight;
-
-    /* health level at which the NPC will surrender */
-    int surrender_level;
-} npc_battle_t;
 
 /* A non-playable character in game */
 typedef struct npc {
@@ -67,7 +33,10 @@ typedef struct npc {
     /* pointer to an existing class struct */
     class_t *class; 
 
-    /* pointer to an existing npc_battle struct */
+    /* boolean representing whether or not the NPC will engage in battles */
+    bool will_fight;
+
+    /* either NULL or a pointer to an existing npc_battle struct */
     npc_battle_t *npc_battle;
 } npc_t;
 
@@ -79,60 +48,6 @@ typedef struct npc npc_hash_t;
 
 
 // STRUCT FUNCTIONS -----------------------------------------------------------
-
-/*
- * Initializes an npc's battle info (npc_battle) with the given parameters.
- *
- * Parameters:
- *  npc_battle: an npc_battle struct; must point to already allocated memory
- *  health: an int with the npc's starting health level
- *  stats: a pointer to an existing stat_t struct defining the npc's battle 
-           stats (see /include/battle/battle_structs.h)
- *  moves: a pointer to an existing move_t struct defining the npc's battle 
-           moves (see /include/battle/battle_structs.h)
- *  ai: the npc's difficulty level (see /include/battle/battle_common.h)
- *  hostility_level: the npc's hostility level 
- *  will_fight: a boolean describing whether or not the npc will fight
- *  surrender_level: the level of health at which the npc surrenders the battle
- *
- * Returns:
- *  SUCCESS on success, FAILURE if an error occurs
- */
-int npc_battle_init(npc_battle_t *npc_battle, int health, stat_t* stats, 
-                    move_t* moves, difficulty_t ai, hostility_t hostility_level,
-                    bool will_fight, int surrender_level);
-/*
- * Allocates a new npc_battle struct in the heap.
- *
- * Parameters:
- *  health: an int with the npc's starting health level
- *  stats: a pointer to an existing stat_t struct defining the npc's battle 
-           stats (see /include/battle/battle_structs.h)
- *  moves: a pointer to an existing move_t struct defining the npc's battle 
-           moves (see /include/battle/battle_structs.h)
- *  ai: the npc's difficulty level (see /include/battle/battle_common.h)
- *  hostility_level: the npc's hostility level 
- *  will_fight: a boolean describing whether or not the npc will fight
- *  surrender_level: the level of health at which the npc surrenders the battle
- *
- * Returns:
- *  pointer to allocated npc_battle
- */
-npc_battle_t *npc_battle_new(int health, stat_t* stats, move_t* moves, 
-		             difficulty_t ai, hostility_t hostility_level, 
-			     bool will_fight, int surrender_level);
-
-/*
- * Frees resources associated with an npc_battle struct.
- *
- * Parameters:
- *  npc_battle: the npc_battle struct to be freed
- *
- * Returns:
- *  SUCCESS if successful, FAILURE if an error occurs
- */
-int npc_battle_free(npc_battle_t *npc_battle);
-
 
 /*
  * Initializes an npc with the given parameters.
@@ -148,14 +63,15 @@ int npc_battle_free(npc_battle_t *npc_battle);
  *             inventory (see /include/game-state/item.h)
  *  class: a pointer to an existing class_t struct defining the npc's class
            (see /include/playerclass/class_structs.h)
- *  npc_battle: a pointer to an existing npc_battle_t struct defining the npc's
- *              battle stats
+ *  will_fight: a boolean describing whether the npc will engage in battle
+ *  npc_battle: either NULL or a pointer to an existing npc_battle_t struct 
+ *              defining the npc's battle stats
  * Returns:
  *  SUCCESS on success, FAILURE if an error occurs
  */
 int npc_init(npc_t *npc, char *npc_id, char *short_desc, char *long_desc,
              convo_t *dialogue, item_hash_t *inventory, class_t *class, 
-             npc_battle_t *npc_battle);
+             bool will_fight, npc_battle_t *npc_battle);
 
 /*
  * Allocates a new npc in the heap.
@@ -170,15 +86,16 @@ int npc_init(npc_t *npc, char *npc_id, char *short_desc, char *long_desc,
  *             inventory (see /include/game-state/item.h)
  *  class: a pointer to an existing class_t struct defining the npc's class
            (see /include/playerclass/class_structs.h)
- *  npc_battle: a pointer to an existing npc_battle_t struct defining the npc's
- *              battle stats
+ *  will_fight: a boolean describing whether the npc will engage in battle
+ *  npc_battle: either NULL or a pointer to an existing npc_battle_t struct 
+ *              defining the npc's battle stats
  *
  * Returns:
  *  pointer to allocated npc
  */
 npc_t *npc_new(char *npc_id, char *short_desc, char *long_desc,
                convo_t *dialogue, item_hash_t *inventory, class_t *class, 
-               npc_battle_t *npc_battle);
+               bool will_fight, npc_battle_t *npc_battle);
 
 /*
  * Frees resources associated with an npc.
