+++ src/npc/src/npc.c
@@ -3,60 +3,9 @@
 // STRUCT FUNCTIONS -----------------------------------------------------------
 
 /* See npc.h */
-int npc_battle_init(npc_battle_t *npc_battle, int health, stat_t* stats,
-                    move_t* moves, difficulty_t ai, hostility_t hostility_level,
-                    bool will_fight, int surrender_level)
-{
-    assert(npc_battle != NULL);
-    npc_battle->health = health;
-    npc_battle->stats = stats;
-    npc_battle->moves = moves;
-    npc_battle->ai = ai;
-    npc_battle->hostility_level = hostility_level;
-    npc_battle->will_fight = will_fight;
-    npc_battle->surrender_level = surrender_level;
-
-    return SUCCESS;
-}
-
-/* See npc.h */
-npc_battle_t *npc_battle_new(int health, stat_t* stats, move_t* moves, 
-		             difficulty_t ai, hostility_t hostility_level, 
-			     bool will_fight, int surrender_level)
-{
-    npc_battle_t *npc_battle;
-    npc_battle = malloc(sizeof(npc_battle_t));
-    memset(npc_battle, 0, sizeof(npc_battle_t));
-    npc_battle->stats = malloc(sizeof(stat_t));
-    npc_battle->moves = malloc(sizeof(move_t)); 
-
-    int check = npc_battle_init(npc_battle, health, stats, moves, ai, 
-                                hostility_level, will_fight, surrender_level);
-
-    if (npc_battle == NULL || npc_battle->stats == NULL ||  
-        npc_battle->moves == NULL || check != SUCCESS)
-    {
-        return NULL;
-    }
-
-    return npc_battle;
-}
-
-/* See npc.h  */
-int npc_battle_free(npc_battle_t *npc_battle)
-{
-    assert(npc_battle != NULL);
-    free(npc_battle->stats); /*waiting for battle team's stat_free function*/
-    move_free(npc_battle->moves);
-    free(npc_battle);
-
-    return SUCCESS;
-}
-
-/* See npc.h */
 int npc_init(npc_t *npc, char *npc_id, char *short_desc, char *long_desc,
              convo_t *dialogue, item_hash_t *inventory, class_t *class,
-             npc_battle_t *npc_battle)
+             bool will_fight, npc_battle_t *npc_battle)
 {
     assert(npc != NULL);
     strcpy(npc->npc_id, npc_id);
@@ -65,6 +14,7 @@ int npc_init(npc_t *npc, char *npc_id, char *short_desc, char *long_desc,
     npc->dialogue = dialogue;
     npc->inventory = inventory;
     npc->class = class;
+    npc->will_fight = will_fight;
     npc->npc_battle = npc_battle;
 
     return SUCCESS;
@@ -73,7 +23,7 @@ int npc_init(npc_t *npc, char *npc_id, char *short_desc, char *long_desc,
 /* See npc.h */
 npc_t *npc_new(char *npc_id, char *short_desc, char *long_desc,
                convo_t *dialogue, item_hash_t *inventory, class_t *class,
-               npc_battle_t *npc_battle)
+               bool will_fight, npc_battle_t *npc_battle)
 {
     npc_t *npc;
     npc = malloc(sizeof(npc_t));
@@ -87,7 +37,7 @@ npc_t *npc_new(char *npc_id, char *short_desc, char *long_desc,
     npc->npc_battle = malloc(sizeof(npc_battle_t));
 
     int check = npc_init(npc, npc_id, short_desc, long_desc, dialogue, inventory
-                         , class, npc_battle); 
+                         , class, will_fight, npc_battle); 
 
     if (npc == NULL || npc->npc_id == NULL ||  npc->short_desc == NULL ||
         npc->long_desc == NULL || npc->dialogue == NULL ||
